ssts_api/requirements.txt
asgiref==3.8.1
click==8.1.7
dj-rest-auth==7.0.0
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
